Refactor Tumor Purity Prediction and Somatic Haplotagging Parameters
- Rename `PurityPrediction` to `PurityEstimation` throughout the codebase for consistency.
- Update parameter handling in `SomaticHaplotag` and related classes to reflect the new naming conventions.
- Modify logging and output messages to ensure clarity in tumor purity estimation results.
- Enhance documentation to accurately describe the updated command usage and parameters for somatic haplotagging and tumor purity estimation.
- Remove deprecated `PurityPrediction` files and update the Makefile accordingly.
