+++ docs/haplotag.md
@@ -1,4 +1,4 @@
-### Haplotag command
+### Haplotagging command
 This command tags (assigns) each read (in BAM) to one haplotype in the phased SNP/SV VCF. i.e., reads will be tagged as HP:i:1 or HP:i:2. In addition, the haplotype block of each read is stored in the PS tag. The phased VCF can be also generated by other programs as long as the PS or HP tags are encoded. The author can specify ```--log``` for additionally output a plain-text file containing haplotype tags of each read without parsing BAM.
 ```
 longphase-s haplotag \
@@ -10,7 +10,7 @@ longphase-s haplotag \
 -o tagged_bam_prefix
 ```
 
-#### The complete list of haplotag parameters
+#### The complete list of haplotagging parameters
 
 ```
 Usage:  haplotag [OPTION] ... READSFILE
