+++ docs/somatic_haplotag.md
@@ -1,18 +1,18 @@
-## Somatic haplotag command
-- [SNP-only somatic haplotag](#snp-only-somatic-haplotag)
+## Somatic haplotagging command
+- [SNP-only somatic haplotaging](#snp-only-somatic-haplotagging)
 - [Somatic haplotagging benchmark](#somatic-haplotagging-benchmark)
-- [The complete list of somatic haplotag parameter](#the-complete-list-of-somatic-haplotag-parameters)
-- [Ouputfiles](#output-files)
-  - [Tumor purity prediction result file](#tumor-purity-prediction-result-file)
+- [The complete list of somatic haplotagging parameters](#the-complete-list-of-somatic-haplotagging-parameters)
+- [Output files](#output-files)
+  - [Tumor purity estimation result file](#tumor-purity-estimation-result-file)
   - [Somatic calling result VCF file](#somatic-calling-result-vcf-file)
   - [Benchmark metrics file](#benchmark-metrics-file)
 
 
-### SNP-only somatic haplotag
-This command performs tumor purity prediction and somatic variant calling, then tags (assigns) each read in tumor BAM to one haplotype based on phased normal SNP VCF and tumor SNP VCF. The reads will be tagged as:
-- HP:z:1 or HP:z:2 for reads with germline SNPs
-- HP:z:H1-1 or HP:z:H2-1 for reads with somatic SNPs derived from germline haplotype 1 or 2
-- HP:z:H3 for reads with somatic SNPs that cannot be derived from germline haplotypes
+### SNP-only somatic haplotagging
+This command performs tumor purity estimation and somatic variant calling using tumor-normal pair BAM files, then tags (assigns) each read in tumor BAM to one haplotype based on phased normal SNP VCF and tumor SNP VCF. The reads will be tagged as:
+- `HP:z:1` or `HP:z:2` for reads with germline SNPs
+- `HP:z:1-1` or `HP:z:2-1` for reads with somatic SNPs derived from germline haplotype 1 or 2
+- `HP:z:3` for reads with somatic SNPs that cannot be derived from germline haplotypes
 
 In addition, the haplotype block of each read is stored in the PS tag (only for reads with phased SNPs). The phased VCF can be generated by other programs as long as the PS or HP tags are encoded. The author can specify `--log` for additionally output a plain-text file containing haplotype tags of each read without parsing BAM.
 
@@ -53,7 +53,7 @@ The benchmark evaluates the performance of somatic haplotagging at the read leve
 The [benchmark results](#benchmark-metrics-file) are output to a metrics file when truth files (`--truth-vcf` and optionally `--truth-bed`) are provided.
 
 
-### The complete list of somatic haplotag parameters
+### The complete list of somatic haplotagging parameters
 ---
 ```
 Usage:  somatic_haplotag [OPTION] ... READSFILE
@@ -82,7 +82,7 @@ optional arguments:
 
 somatic variant calling arguments:
       --tumor-purity=Num              tumor purity (0.1~1.0) for adjusting somatic variant filtering.
-                                      lower values apply more conservative filters. default: auto-prediction.
+                                      lower values apply more conservative filters. default: automatic estimation.
       --disableFilter                 disable somatic variant filtering and accept all tumor VCF variants. default: false.
       --output-somatic-vcf            output filtered somatic variants VCF file based on tumor VCF. default: false.
                                       variants passing filters will be marked as PASS, others as LowQual.
@@ -98,15 +98,15 @@ somatic haplotagging benchmark arguments:
 
 ### Output files
 ---
-### Tumor purity prediction result file
-The tumor purity prediction reuslt file contains the estimated tumor purity value and related statistics. The file includes:
+### Tumor purity estimation result file
+The tumor purity estimation result file contains the estimated tumor purity value and related statistics. The file includes:
 - Tumor purity: The estimated tumor purity value (0.0-1.0)
-- Filter parameters: Parameters used for filtering the prediction data
-- Data size: Number of variants used for prediction
+- Filter parameters: Parameters used for filtering the estimation data
+- Data size: Number of variants used for estimation
 
-An example of tumor purity prediction file
+An example of tumor purity estimation file
 ```
-#==========Prediction result===========
+#==========Estimation result===========
 Tumor purity: 0.169599
 ```
 
@@ -125,7 +125,7 @@ All other information in the VCF file remains unchanged from the input tumor VCF
 
 ```
 ##longphase_s_version=1.0.0
-##commandline=longphase_s somatic_haplotag -s normal_phased.vcf -b normal.bam --tumor-snp-file 
+##commandline=longphase_s somatic_haplotag -s phased_normal.vcf -b normal.bam --tumor-snp-file 
 tumor.vcf --tumor-bam-file tumor.bam -r GRCh38_no_alt_analysis_set.fasta -t 64 -o tagged_tumor_bam --tagSupplementary -q 20 --output-somatic-vcf
 #CHROM  POS     ID      REF     ALT     QUAL    FILTER  INFO    FORMAT  SAMPLE
 ```
@@ -155,11 +155,11 @@ An example of benchmark metrics file
 ##Truth VCF File: truth.vcf
 ##Truth BED File: truth.bed
 ##MappingQualityThreshold: 20
-##Tatal reads: 8804601
-##Tatal truth somatic reads: 149518
-##Tatal truth HP1-1: 69793
-##Tatal truth HP2-1: 68475
-##Tatal truth HP3: 11250
+##Total reads: 8804601
+##Total truth somatic reads: 149518
+##Total truth HP1-1: 69793
+##Total truth HP2-1: 68475
+##Total truth HP3: 11250
 ## Haplotype   Precision      Recall         F1-Score       TP             FP             FN             
 ##-----------------------------------------------------------------------------------------------
 ## HP1-1       0.9840         0.6343         0.7714         44269          721            25524
